pyproject.toml
[project]
name = "multimodal-stock-prediction"
version = "1.0.0"
description = "Multi-Modal Gated Transformer for Japanese Stock Return Prediction using Financial Text and Market Data"
readme = "README.md"
requires-python = ">=3.12,<3.14"
license = { text = "MIT" }
authors = [{ name = "okima" }]

dependencies = [
    # Core
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.2.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "joblib>=1.3.0",
    # Deep Learning
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "datasets>=2.14.0",
    "accelerate>=0.20.0",
    "safetensors>=0.4.0",
    # Machine Learning
    "lightgbm>=4.0.0",
    "optuna>=3.0.0",
    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "japanize-matplotlib>=1.1.0",
    # Finance
    "yfinance>=0.2.0",
    "pandas-ta>=0.3.14b1",
    # NLP / Text Processing
    "beautifulsoup4>=4.12.0",
    "feedparser>=6.0.0",
    "fugashi>=1.2.0",
    "ipadic>=1.0.0",
    "python-dateutil>=2.8.2",
    # API
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "pytest>=8.0.0",
    "pre-commit>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["2_src"]

[tool.uv]
# PyTorch CUDA 12.1 index for GPU support
index-strategy = "unsafe-best-match"
extra-index-url = ["https://download.pytorch.org/whl/cu121"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "E402",  # module level import not at top (common in scripts)
    "E701",  # multiple statements on one line (common in research scripts)
    "E722",  # bare except (existing research code pattern)
    "F841",  # local variable assigned but never used
    "B007",  # unused loop control variable
    "B006",  # mutable argument default (existing research code)
    "B023",  # function uses loop variable
    "B905",  # zip without strict
    "W293",  # blank line with whitespace in multiline strings
]

[tool.ruff.lint.isort]
known-first-party = ["arch", "layers"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["2_src", "2_src/models"]
